Fix desktop sync silently dropping/failing offline writes, add real-time sync and reactive PK-conflict resolution
Root cause of "20 offline records, all failed to sync" wasn't PK
collisions - it was two endpoints never wired for push-back at all:
Suppliers (so any Bill created against an offline-added supplier was
permanently un-syncable, 404ing forever against production) and the
receiving flow (bill_receive_items/bill_confirm_received - "receiving
from supplier" - silently never even attempted to sync).

Also found two real, previously-live overflow bugs while testing the
actual reserved-PK-range replay path end-to-end for the first time:
NumberSequence.next_value and several loose "reference_id" columns
(SupplierLedger, CustomerLedger, RecordDeletionLog, StockMovement,
StockReservation, InventoryLock, PurchaseApproval) were plain
IntegerField/PositiveIntegerField, which can't hold a value from a
desktop device's reserved range (floor 10,000,000,000) - Postgres's
plain integer maxes out around 2.1B. A real leftover production row
from earlier device-range testing already overflowed one of these the
moment its counter was first seeded, confirming this wasn't
theoretical.

Changes:
- Suppliers, Products, Expenses, Customer edits, sales returns, and
  the bill-receiving flow now have the same client-supplied-PK
  acceptance the 5 original endpoints had (core/device_registry.py's
  validated_desktop_pk/validated_desktop_number).
- Device re-pairing now reuses its own previously-registered PK range
  instead of always minting a new one (core/models.py's new
  DesktopDeviceIdentity), which was silently orphaning anything still
  queued from before a re-pair.
- A genuine PK collision (rare now, but possible) self-heals in the
  same request via a SAVEPOINT retry (core/pk_conflict.py) and reports
  back so the desktop can renumber its local copy to match
  (core/desktop_pk_renumber.py) - scoped to the known Phase-C model
  set, not a blind whole-schema walk.
- The desktop app now triggers a push attempt immediately after each
  offline write when reachable, instead of waiting up to 5 minutes or
  a manual Sync Now click.
- Fixed a real bug in the fix itself: SQLite's PRAGMA foreign_keys is
  a no-op once already inside a transaction, so the renumber
  transaction (and core/snapshot.py's existing import, which had the
  identical bug) needed constraint_checks_disabled() to wrap
  transaction.atomic(), not nest inside it.

All verified against real production (rolled-back transactions where
possible, explicit cleanup otherwise) - not just "the code compiles."

+++ core/models.py
@@ -65,7 +65,12 @@ class RecordDeletionLog(models.Model):
 
     company = models.ForeignKey('user_auth.Company', on_delete=models.CASCADE, related_name='deletion_logs')
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
+    # PositiveBigIntegerField, not PositiveIntegerField: this audit log covers every
+    # soft-deletable model, including desktop-created rows whose PK falls inside a
+    # reserved device range (core/device_registry.py, floor 10_000_000_000) - already
+    # past Postgres's plain `integer` max (~2.1B). See purchase.SupplierLedger.
+    # reference_id's matching comment for where this bug class was first hit.
+    object_id = models.PositiveBigIntegerField()
     content_object = GenericForeignKey('content_type', 'object_id')
     object_repr = models.CharField(max_length=255)
     action = models.CharField(max_length=10, choices=ACTION_CHOICES)
@@ -94,5 +99,5 @@ class RecordDeletionLog(models.Model):
 # class is defined would deadlock the circular import.
 from core.idempotency import IdempotencyKey  # noqa: E402,F401
 from core.numbering import NumberSequence  # noqa: E402,F401
-from core.device_registry import DeviceIdRangeCounter, DeviceRegistration  # noqa: E402,F401
+from core.device_registry import DeviceIdRangeCounter, DeviceRegistration, DesktopDeviceIdentity  # noqa: E402,F401
 from core.desktop_sync_queue import DesktopSyncQueueEntry  # noqa: E402,F401
